Create Porter Delivery Time Estimation
Porter has emerged out to be the one-stop solution when it comes to 
logistics management, irrespective of the sphere whether it is domestic 
or commercial.  
When it comes to logistics handling, safety & delivery time are a couple 
of things that strikes at the very first instance. 
In this project, we'll estimate the time taken for food delivery on the 
basis of a number of variables which plays a significant role in customer 
& the firm's decision making.